Added all the SCHOOL types (and converted to 64bit as it was needed with all the new types.
Also made action_t derive from active_spell_t and provided some constructors so you can choose to specify a spell name or id rather than using the old-style method.

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@4765 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_item.cpp
@@ -566,67 +566,67 @@ bool item_t::decode_special( special_effect_t& effect,
     {
       effect.trigger_str  = t.full;
       effect.trigger_type = PROC_DAMAGE;
-      effect.trigger_mask = (1 << SCHOOL_ARCANE);
+      effect.trigger_mask = ( int64_t( 1 ) << SCHOOL_ARCANE);
     }
     else if ( t.full == "onbleeddamage" )
     {
       effect.trigger_str  = t.full;
       effect.trigger_type = PROC_DAMAGE;
-      effect.trigger_mask = (1 << SCHOOL_BLEED);
+      effect.trigger_mask = ( int64_t( 1 ) << SCHOOL_BLEED);
     }
     else if ( t.full == "onchaosdamage" )
     {
       effect.trigger_str  = t.full;
       effect.trigger_type = PROC_DAMAGE;
-      effect.trigger_mask = (1 << SCHOOL_CHAOS);
+      effect.trigger_mask = ( int64_t( 1 ) << SCHOOL_CHAOS);
     }
     else if ( t.full == "onfiredamage" )
     {
       effect.trigger_str  = t.full;
       effect.trigger_type = PROC_DAMAGE;
-      effect.trigger_mask = (1 << SCHOOL_FIRE);
+      effect.trigger_mask = ( int64_t( 1 ) << SCHOOL_FIRE);
     }
     else if ( t.full == "onfrostdamage" )
     {
       effect.trigger_str  = t.full;
       effect.trigger_type = PROC_DAMAGE;
-      effect.trigger_mask = (1 << SCHOOL_FROST);
+      effect.trigger_mask = ( int64_t( 1 ) << SCHOOL_FROST);
     }
     else if ( t.full == "onfrostfiredamage" )
     {
       effect.trigger_str  = t.full;
       effect.trigger_type = PROC_DAMAGE;
-      effect.trigger_mask = (1 << SCHOOL_FROSTFIRE);
+      effect.trigger_mask = ( int64_t( 1 ) << SCHOOL_FROSTFIRE);
     }
     else if ( t.full == "onholydamage" )
     {
       effect.trigger_str  = t.full;
       effect.trigger_type = PROC_DAMAGE;
-      effect.trigger_mask = (1 << SCHOOL_HOLY);
+      effect.trigger_mask = ( int64_t( 1 ) << SCHOOL_HOLY);
     }
     else if ( t.full == "onnaturedamage" )
     {
       effect.trigger_str  = t.full;
       effect.trigger_type = PROC_DAMAGE;
-      effect.trigger_mask = (1 << SCHOOL_NATURE);
+      effect.trigger_mask = ( int64_t( 1 ) << SCHOOL_NATURE);
     }
     else if ( t.full == "onphysicaldamage" )
     {
       effect.trigger_str  = t.full;
       effect.trigger_type = PROC_DAMAGE;
-      effect.trigger_mask = (1 << SCHOOL_PHYSICAL);
+      effect.trigger_mask = ( int64_t( 1 ) << SCHOOL_PHYSICAL);
     }
     else if ( t.full == "onshadowdamage" )
     {
       effect.trigger_str  = t.full;
       effect.trigger_type = PROC_DAMAGE;
-      effect.trigger_mask = (1 << SCHOOL_SHADOW);
+      effect.trigger_mask = ( int64_t( 1 ) << SCHOOL_SHADOW);
     }
     else if ( t.full == "ondraindamage" )
     {
       effect.trigger_str  = t.full;
       effect.trigger_type = PROC_DAMAGE;
-      effect.trigger_mask = (1 << SCHOOL_DRAIN);
+      effect.trigger_mask = ( int64_t( 1 ) << SCHOOL_DRAIN);
     }
     else if ( t.full == "ontick" )
     {
